Refactor code structure (see below for details)
### 1. Reorganize layers

Topmost level (TCP server, http server) and language-specific layer
(Mongo, SQL) are reorganized.

The upper level is called `cortex`, which means bark of a tree, i.e.,
the outer layer of a trunk

The http service is named `Unicus` (which means "unique", a derivative
from the word "Unum"), and serves as a paralell interface along with
Mongo and MySQL.

The Unicus and Mongo TCP handlers are extracted from `main.rs`.

### 2. Reorganize Mongo parser

The type declarations, parsers, and (in the future) serializers are
organized by `op`s. For instance, for `OP_QUERY`, are its relevant code
is moved to `src/cortices/mongo/ops/op_query.rs`.

### 3. Centralized configuration

Configuration code (ports of servers, choice of KV engine, etc.) are
centralized into `src/config.rs`

### 4. Error message cleanup

Some of error messages deep in the functions is remove; they should be
handled by a global error reporter that operates on UnumError, rather
than scattered around the codebase